+++ src/main/kotlin/net/pwall/json/test/JSONExpect.kt
@@ -25,10 +25,13 @@
 
 package net.pwall.json.test
 
+import kotlin.reflect.KClass
+import kotlin.reflect.KType
 import kotlin.test.fail
 
 import java.math.BigDecimal
 
+import net.pwall.json.JSONTypeRef
 import net.pwall.json.parseJSON
 
 /**
@@ -48,7 +51,7 @@ class JSONExpect private constructor(private val obj: Any?, private val pathInfo
         if (obj !is Int)
             failOnType("integer")
         if (obj != expected)
-            failOnValue(expected.toString(), obj.toString())
+            failOnValue(expected, obj)
     }
 
     /**
@@ -61,7 +64,7 @@ class JSONExpect private constructor(private val obj: Any?, private val pathInfo
         if (obj !is Long)
             failOnType("long integer")
         if (obj != expected)
-            failOnValue(expected.toString(), obj.toString())
+            failOnValue(expected, obj)
     }
 
     /**
@@ -74,7 +77,7 @@ class JSONExpect private constructor(private val obj: Any?, private val pathInfo
         if (obj !is BigDecimal)
             failOnType("decimal")
         if (obj != expected)
-            failOnValue(expected.toString(), obj.toString())
+            failOnValue(expected, obj)
     }
 
     /**
@@ -87,7 +90,7 @@ class JSONExpect private constructor(private val obj: Any?, private val pathInfo
         if (obj !is Boolean)
             failOnType("boolean")
         if (obj != expected)
-            failOnValue(expected.toString(), obj.toString())
+            failOnValue(expected, obj)
     }
 
     /**
@@ -109,6 +112,129 @@ class JSONExpect private constructor(private val obj: Any?, private val pathInfo
                     failOnValue("\"$expected\"", "\"$obj\"")
             }
         }
+    }
+
+    /**
+     * Check the value as a member of an [IntRange].
+     *
+     * @param   expected        the [IntRange]
+     * @throws  AssertionError  if the value is not within the [IntRange]
+     */
+    fun value(expected: IntRange) {
+        if (obj !is Int)
+            failOnType("integer")
+        if (obj !in expected)
+            failInRange(obj)
+    }
+
+    /**
+     * Check the value as a member of a [LongRange].
+     *
+     * @param   expected        the [LongRange]
+     * @throws  AssertionError  if the value is not within the [LongRange]
+     */
+    fun value(expected: LongRange) {
+        if (obj !is Long)
+            failOnType("long integer")
+        if (obj !in expected)
+            failInRange(obj)
+    }
+
+    /**
+     * Check the value as a member of a [ClosedRange].  This will normally be invoked via the inline function.
+     *
+     * @param   expected        the [ClosedRange]
+     * @param   type            the type of the elements of the [ClosedRange]
+     * @throws  AssertionError  if the value is not within the [ClosedRange]
+     */
+    @Suppress("UNCHECKED_CAST")
+    fun <T: Comparable<T>> valueInRange(expected: ClosedRange<T>, type: KType) {
+        val itemClass = type.classifier as? KClass<*> ?: fail("${prefix}Can't determine type of ClosedRange")
+        when (itemClass) {
+            Int::class -> {
+                if (obj !is Int)
+                    failOnType("integer")
+                if (obj as T !in expected)
+                    failInRange(obj)
+            }
+            Long::class -> {
+                if (obj !is Long)
+                    failOnType("long integer")
+                if (obj as T !in expected)
+                    failInRange(obj)
+            }
+            BigDecimal::class -> {
+                if (obj !is BigDecimal)
+                    failOnType("decimal")
+                if (obj as T !in expected)
+                    failInRange(obj)
+            }
+            String::class -> {
+                if (obj !is String)
+                    failOnType("string")
+                if (obj as T !in expected)
+                    failInRange("\"$obj\"")
+            }
+            else -> fail("${prefix}Can't perform test using range of $type")
+        }
+    }
+
+    /**
+     * Check the value as a member of a [ClosedRange].
+     *
+     * @param   expected        the [ClosedRange]
+     * @throws  AssertionError  if the value is not within the [ClosedRange]
+     */
+    inline fun <reified T: Comparable<T>> value(expected: ClosedRange<T>) {
+        valueInRange(expected, JSONTypeRef.create<T>(nullable = true).refType)
+    }
+
+    /**
+     * Check the value as a member of a [Collection].  This will normally be invoked via the inline function.
+     *
+     * @param   expected        the [Collection]
+     * @param   type            the type of the elements of the [Collection]
+     * @throws  AssertionError  if the value does not match any element of the [Collection]
+     */
+    fun <T> valueInCollection(expected: Collection<T>, type: KType) {
+        val itemClass = type.classifier as? KClass<*> ?: fail("${prefix}Can't determine type of Collection")
+        when (itemClass) {
+            Int::class -> {
+                if (obj != null && obj !is Int)
+                    failOnType("integer")
+                if (!expected.contains(obj))
+                    failInCollection(obj)
+            }
+            Long::class -> {
+                if (obj != null && obj !is Long)
+                    failOnType("long integer")
+                if (!expected.contains(obj))
+                    failInCollection(obj)
+            }
+            BigDecimal::class -> {
+                if (obj != null && obj !is BigDecimal)
+                    failOnType("decimal")
+                if (!expected.contains(obj))
+                    failInCollection(obj)
+            }
+            String::class -> {
+                if (obj != null && obj !is String)
+                    failOnType("string")
+                if (!expected.contains(obj))
+                    failInCollection(if (obj == null) "null" else "\"$obj\"")
+            }
+            else -> fail("${prefix}Can't perform test using collection of $type")
+        }
+    }
+
+    /**
+     * Check the value as a member of a [Collection].
+     *
+     * @param   expected        the [Collection]
+     * @throws  AssertionError  if the value does not match any element of the [Collection]
+     */
+    inline fun <reified T: Any> value(expected: Collection<T?>) {
+        valueInCollection(expected, JSONTypeRef.create<T>(nullable = true).refType)
     }
 
     /**
@@ -174,6 +300,82 @@ class JSONExpect private constructor(private val obj: Any?, private val pathInfo
         checkName(name).let {
             JSONExpect(getProperty(it), propertyPath(it)).value(expected)
         }
+    }
+
+    /**
+     * Check a property as a member of an [IntRange].
+     *
+     * @param   name            the property name
+     * @param   expected        the [IntRange]
+     * @throws  AssertionError  if the value is not within the [IntRange]
+     */
+    fun property(name: String, expected: IntRange) {
+        checkName(name).let {
+            JSONExpect(getProperty(it), propertyPath(it)).value(expected)
+        }
+    }
+
+    /**
+     * Check a property as a member of a [LongRange].
+     *
+     * @param   name            the property name
+     * @param   expected        the [LongRange]
+     * @throws  AssertionError  if the value is not within the [LongRange]
+     */
+    fun property(name: String, expected: LongRange) {
+        checkName(name).let {
+            JSONExpect(getProperty(it), propertyPath(it)).value(expected)
+        }
+    }
+
+    /**
+     * Check a property as a member of a [ClosedRange].  This will normally be invoked via the inline function.
+     *
+     * @param   name            the property name
+     * @param   expected        the [ClosedRange]
+     * @param   type            the type of the elements of the [ClosedRange]
+     * @throws  AssertionError  if the value is not within the [ClosedRange]
+     */
+    fun <T: Comparable<T>> propertyInRange(name: String, expected: ClosedRange<T>, type: KType) {
+        checkName(name).let {
+            JSONExpect(getProperty(it), propertyPath(it)).valueInRange(expected, type)
+        }
+    }
+
+    /**
+     * Check a property as a member of a [ClosedRange].
+     *
+     * @param   name            the property name
+     * @param   expected        the [ClosedRange]
+     * @throws  AssertionError  if the value is not within the [ClosedRange]
+     */
+    inline fun <reified T: Comparable<T>> property(name: String, expected: ClosedRange<T>) {
+        propertyInRange(name, expected, JSONTypeRef.create<T>(nullable = true).refType)
+    }
+
+    /**
+     * Check a property as a member of a [Collection].  This will normally be invoked via the inline function.
+     *
+     * @param   name            the property name
+     * @param   expected        the [Collection]
+     * @param   type            the type of the elements of the [Collection]
+     * @throws  AssertionError  if the value does not match any element of the [Collection]
+     */
+    fun <T> propertyInCollection(name: String, expected: Collection<T>, type: KType) {
+        checkName(name).let {
+            JSONExpect(getProperty(it), propertyPath(it)).valueInCollection(expected, type)
+        }
+    }
+
+    /**
+     * Check a property as a member of a [Collection].
+     *
+     * @param   name            the property name
+     * @param   expected        the [Collection]
+     * @throws  AssertionError  if the value does not match any element of the [Collection]
+     */
+    inline fun <reified T: Any> property(name: String, expected: Collection<T?>) {
+        propertyInCollection(name, expected, JSONTypeRef.create<T>(nullable = true).refType)
     }
 
     /**
@@ -255,6 +457,82 @@ class JSONExpect private constructor(private val obj: Any?, private val pathInfo
     }
 
     /**
+     * Check an array item as a member of an [IntRange].
+     *
+     * @param   index           the array index
+     * @param   expected        the [IntRange]
+     * @throws  AssertionError  if the value is not within the [IntRange]
+     */
+    fun item(index: Int, expected: IntRange) {
+        checkIndex(index).let {
+            JSONExpect(getItem(it), itemPath(it)).value(expected)
+        }
+    }
+
+    /**
+     * Check an array item as a member of a [LongRange].
+     *
+     * @param   index           the array index
+     * @param   expected        the [LongRange]
+     * @throws  AssertionError  if the value is not within the [LongRange]
+     */
+    fun item(index: Int, expected: LongRange) {
+        checkIndex(index).let {
+            JSONExpect(getItem(it), itemPath(it)).value(expected)
+        }
+    }
+
+    /**
+     * Check an array item as a member of a [ClosedRange].  This will normally be invoked via the inline function.
+     *
+     * @param   index           the array index
+     * @param   expected        the [ClosedRange]
+     * @param   type            the type of the elements of the [ClosedRange]
+     * @throws  AssertionError  if the value is not within the [ClosedRange]
+     */
+    fun <T: Comparable<T>> itemInRange(index: Int, expected: ClosedRange<T>, type: KType) {
+        checkIndex(index).let {
+            JSONExpect(getItem(it), itemPath(it)).valueInRange(expected, type)
+        }
+    }
+
+    /**
+     * Check an array item as a member of a [ClosedRange].
+     *
+     * @param   index           the array index
+     * @param   expected        the [ClosedRange]
+     * @throws  AssertionError  if the value is not within the [ClosedRange]
+     */
+    inline fun <reified T: Comparable<T>> item(index: Int, expected: ClosedRange<T>) {
+        itemInRange(index, expected, JSONTypeRef.create<T>(nullable = true).refType)
+    }
+
+    /**
+     * Check an array item as a member of a [Collection].  This will normally be invoked via the inline function.
+     *
+     * @param   index           the array index
+     * @param   expected        the [Collection]
+     * @param   type            the type of the elements of the [Collection]
+     * @throws  AssertionError  if the value does not match any element of the [Collection]
+     */
+    fun <T> itemInCollection(index: Int, expected: Collection<T>, type: KType) {
+        checkIndex(index).let {
+            JSONExpect(getItem(it), itemPath(it)).valueInCollection(expected, type)
+        }
+    }
+
+    /**
+     * Check an array item as a member of a [Collection].
+     *
+     * @param   index           the array index
+     * @param   expected        the [Collection]
+     * @throws  AssertionError  if the value does not match any element of the [Collection]
+     */
+    inline fun <reified T: Any> item(index: Int, expected: Collection<T?>) {
+        itemInCollection(index, expected, JSONTypeRef.create<T>(nullable = true).refType)
+    }
+
+    /**
      * Select an array item for nested tests.
      *
      * @param   index           the array index
@@ -298,8 +576,8 @@ class JSONExpect private constructor(private val obj: Any?, private val pathInfo
             fail("${prefix}JSON property not absent - $name")
     }
 
-    private fun failOnValue(expectedString: String, actualString: String): Nothing {
-        fail("${prefix}JSON value doesn't match - Expected $expectedString, was $actualString")
+    private fun failOnValue(expected: Any, actual: Any): Nothing {
+        fail("${prefix}JSON value doesn't match - Expected $expected, was $actual")
     }
 
     private fun failOnType(expected: String): Nothing {
@@ -315,6 +593,14 @@ class JSONExpect private constructor(private val obj: Any?, private val pathInfo
             else -> "unknown"
         }
         fail("${prefix}JSON type doesn't match - Expected $expected, was $type")
+    }
+
+    private fun failInCollection(actual: Any?) {
+        fail("${prefix}JSON value not in collection - $actual")
+    }
+
+    private fun failInRange(actual: Any?) {
+        fail("${prefix}JSON value not in range - $actual")
     }
 
     private fun checkName(name: String): String =