+++ src/test/kotlin/net/pwall/json/test/JSONExpectTest.kt
@@ -610,4 +610,459 @@ class JSONExpectTest {
         expect("abc: JSON type doesn't match - Expected string, was integer") { exception.message }
     }
 
+    @Test fun `should test int value as member of a collection`() {
+        val json = "123"
+        expectJSON(json) {
+            value(setOf(123, 456, 789))
+        }
+    }
+
+    @Test fun `should fail on incorrect test of int value as member of a collection`() {
+        val json = "124"
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                value(setOf(123, 456, 789))
+            }
+        }
+        expect("JSON value not in collection - 124") { exception.message }
+    }
+
+    @Test fun `should test null as member of a collection of int`() {
+        val json = "null"
+        expectJSON(json) {
+            value(setOf(123, 456, 789, null))
+        }
+    }
+
+    @Test fun `should fail on incorrect test of null as member of a collection of int`() {
+        val json = "null"
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                value(setOf(123, 456, 789))
+            }
+        }
+        expect("JSON value not in collection - null") { exception.message }
+    }
+
+    @Test fun `should test long value as member of a collection`() {
+        val json = "123456789123456789"
+        expectJSON(json) {
+            value(setOf(123456789123456789, 0L))
+        }
+    }
+
+    @Test fun `should fail on incorrect test of long value as member of a collection`() {
+        val json = "123456789123456788"
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                value(setOf(123456789123456789, 0L))
+            }
+        }
+        expect("JSON value not in collection - 123456789123456788") { exception.message }
+    }
+
+    @Test fun `should test null as member of a collection of long`() {
+        val json = "null"
+        expectJSON(json) {
+            value(setOf(123456789123456789, 0L, null))
+        }
+    }
+
+    @Test fun `should fail on incorrect test of null as member of a collection of long`() {
+        val json = "null"
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                value(setOf(123456789123456789, 0L))
+            }
+        }
+        expect("JSON value not in collection - null") { exception.message }
+    }
+
+    @Test fun `should test decimal value as member of a collection`() {
+        val json = "9.99"
+        expectJSON(json) {
+            value(setOf(BigDecimal("9.99"), BigDecimal("19.99")))
+        }
+    }
+
+    @Test fun `should fail on incorrect test of decimal value as member of a collection`() {
+        val json = "29.99"
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                value(setOf(BigDecimal("9.99"), BigDecimal("19.99")))
+            }
+        }
+        expect("JSON value not in collection - 29.99") { exception.message }
+    }
+
+    @Test fun `should test null as member of a collection of decimal`() {
+        val json = "null"
+        expectJSON(json) {
+            value(setOf(BigDecimal("9.99"), BigDecimal("19.99"), null))
+        }
+    }
+
+    @Test fun `should fail on incorrect test of null as member of a collection of decimal`() {
+        val json = "null"
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                value(setOf(BigDecimal("9.99"), BigDecimal("19.99")))
+            }
+        }
+        expect("JSON value not in collection - null") { exception.message }
+    }
+
+    @Test fun `should test string as member of a collection`() {
+        val json = "\"abc\""
+        expectJSON(json) {
+            value(setOf("abc", "def", "ghi"))
+        }
+    }
+
+    @Test fun `should fail on incorrect test of string as member of a collection`() {
+        val json = "\"abcd\""
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                value(setOf("abc", "def", "ghi"))
+            }
+        }
+        expect("JSON value not in collection - \"abcd\"") { exception.message }
+    }
+
+    @Test fun `should test null as member of a collection of string`() {
+        val json = "null"
+        expectJSON(json) {
+            value(setOf("abc", "def", "ghi", null))
+        }
+    }
+
+    @Test fun `should fail on incorrect test of null as member of a collection of string`() {
+        val json = "null"
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                value(setOf("abc", "def", "ghi"))
+            }
+        }
+        expect("JSON value not in collection - null") { exception.message }
+    }
+
+    @Test fun `should test int property as member of a collection`() {
+        val json = """{"abc":123}"""
+        expectJSON(json) {
+            property("abc", setOf(123, 456, 789))
+        }
+    }
+
+    @Test fun `should fail on incorrect test of int property as member of a collection`() {
+        val json = """{"abc":124}"""
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                property("abc", setOf(123, 456, 789))
+            }
+        }
+        expect("abc: JSON value not in collection - 124") { exception.message }
+    }
+
+    @Test fun `should test null property as member of a collection of int`() {
+        val json = """{"abc":null}"""
+        expectJSON(json) {
+            property("abc", setOf(123, 456, 789, null))
+        }
+    }
+
+    @Test fun `should fail on incorrect test of null property as member of a collection of int`() {
+        val json = """{"abc":null}"""
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                property("abc", setOf(123, 456, 789))
+            }
+        }
+        expect("abc: JSON value not in collection - null") { exception.message }
+    }
+
+    @Test fun `should test string property as member of a collection`() {
+        val json = """{"prop":"abc"}"""
+        expectJSON(json) {
+            property("prop", setOf("abc", "def", "ghi"))
+        }
+    }
+
+    @Test fun `should fail on incorrect test of string property as member of a collection`() {
+        val json = """{"prop":"abcd"}"""
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                property("prop", setOf("abc", "def", "ghi"))
+            }
+        }
+        expect("prop: JSON value not in collection - \"abcd\"") { exception.message }
+    }
+
+    @Test fun `should test null property as member of a collection of string`() {
+        val json = """{"prop":null}"""
+        expectJSON(json) {
+            property("prop", setOf("abc", "def", "ghi", null))
+        }
+    }
+
+    @Test fun `should fail on incorrect test of null property as member of a collection of string`() {
+        val json = """{"prop":null}"""
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                property("prop", setOf("abc", "def", "ghi"))
+            }
+        }
+        expect("prop: JSON value not in collection - null") { exception.message }
+    }
+
+    @Test fun `should test int array item as member of a collection`() {
+        val json = "[123]"
+        expectJSON(json) {
+            item(0, setOf(123, 456, 789))
+        }
+    }
+
+    @Test fun `should fail on incorrect test of int array item as member of a collection`() {
+        val json = "[124]"
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                item(0, setOf(123, 456, 789))
+            }
+        }
+        expect("[0]: JSON value not in collection - 124") { exception.message }
+    }
+
+    @Test fun `should test null array item as member of a collection of int`() {
+        val json = "[null]"
+        expectJSON(json) {
+            item(0, setOf(123, 456, 789, null))
+        }
+    }
+
+    @Test fun `should fail on incorrect test of null array item as member of a collection of int`() {
+        val json = "[null]"
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                item(0, setOf(123, 456, 789))
+            }
+        }
+        expect("[0]: JSON value not in collection - null") { exception.message }
+    }
+
+    @Test fun `should test string array item as member of a collection`() {
+        val json = """["abc"]"""
+        expectJSON(json) {
+            item(0, setOf("abc", "def", "ghi"))
+        }
+    }
+
+    @Test fun `should fail on incorrect test of string array item as member of a collection`() {
+        val json = """["abcd"]"""
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                item(0, setOf("abc", "def", "ghi"))
+            }
+        }
+        expect("[0]: JSON value not in collection - \"abcd\"") { exception.message }
+    }
+
+    @Test fun `should test null array item as member of a collection of string`() {
+        val json = "[null]"
+        expectJSON(json) {
+            item(0, setOf("abc", "def", "ghi", null))
+        }
+    }
+
+    @Test fun `should fail on incorrect test of null array item as member of a collection of string`() {
+        val json = "[null]"
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                item(0, setOf("abc", "def", "ghi"))
+            }
+        }
+        expect("[0]: JSON value not in collection - null") { exception.message }
+    }
+
+    @Test fun `should test int value as member of a range`() {
+        val json = "123"
+        expectJSON(json) {
+            value(120..125)
+        }
+    }
+
+    @Test fun `should fail on incorrect test of int value as member of a range`() {
+        val json = "126"
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                value(120..125)
+            }
+        }
+        expect("JSON value not in range - 126") { exception.message }
+    }
+
+    @Test fun `should test int property as member of a range`() {
+        val json = """{"abc":123}"""
+        expectJSON(json) {
+            property("abc", 120..125)
+        }
+    }
+
+    @Test fun `should fail on incorrect test of int property as member of a range`() {
+        val json = """{"abc":126}"""
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                property("abc", 120..125)
+            }
+        }
+        expect("abc: JSON value not in range - 126") { exception.message }
+    }
+
+    @Test fun `should test int array item as member of a range`() {
+        val json = "[123]"
+        expectJSON(json) {
+            item(0, 120..125)
+        }
+    }
+
+    @Test fun `should fail on incorrect test of int array item as member of a range`() {
+        val json = "[126]"
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                item(0, 120..125)
+            }
+        }
+        expect("[0]: JSON value not in range - 126") { exception.message }
+    }
+
+    @Test fun `should test long value as member of a range`() {
+        val json = "123456789123456789"
+        expectJSON(json) {
+            value(123456789123456780..123456789123456789)
+        }
+    }
+
+    @Test fun `should fail on incorrect test of long value as member of a range`() {
+        val json = "123456789123456779"
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                value(123456789123456780..123456789123456789)
+            }
+        }
+        expect("JSON value not in range - 123456789123456779") { exception.message }
+    }
+
+    @Test fun `should test long property as member of a range`() {
+        val json = """{"abc":123456789123456789}"""
+        expectJSON(json) {
+            property("abc", 123456789123456780..123456789123456789)
+        }
+    }
+
+    @Test fun `should fail on incorrect test of long property as member of a range`() {
+        val json = """{"abc":123456789123456779}"""
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                property("abc", 123456789123456780..123456789123456789)
+            }
+        }
+        expect("abc: JSON value not in range - 123456789123456779") { exception.message }
+    }
+
+    @Test fun `should test long array item as member of a range`() {
+        val json = "[123456789123456789]"
+        expectJSON(json) {
+            item(0, 123456789123456780..123456789123456789)
+        }
+    }
+
+    @Test fun `should fail on incorrect test of long array item as member of a range`() {
+        val json = "[123456789123456779]"
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                item(0, 123456789123456780..123456789123456789)
+            }
+        }
+        expect("[0]: JSON value not in range - 123456789123456779") { exception.message }
+    }
+
+    @Test fun `should test string value as member of a range`() {
+        val json = "\"abcde\""
+        expectJSON(json) {
+            value("abcda".."abcdz")
+        }
+    }
+
+    @Test fun `should fail on incorrect test of string value as member of a range`() {
+        val json = "\"abcde\""
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                value("abcdg".."abcdz")
+            }
+        }
+        expect("JSON value not in range - \"abcde\"") { exception.message }
+    }
+
+    @Test fun `should test string property as member of a range`() {
+        val json = """{"prop":"abcde"}"""
+        expectJSON(json) {
+            property("prop", "abcda".."abcdz")
+        }
+    }
+
+    @Test fun `should fail on incorrect test of string property as member of a range`() {
+        val json = """{"prop":"abcde"}"""
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                property("prop", "abcdg".."abcdz")
+            }
+        }
+        expect("prop: JSON value not in range - \"abcde\"") { exception.message }
+    }
+
+    @Test fun `should test string array item as member of a range`() {
+        val json = "[\"abcde\"]"
+        expectJSON(json) {
+            item(0, "abcda".."abcdz")
+        }
+    }
+
+    @Test fun `should fail on incorrect test of string array item as member of a range`() {
+        val json = """["abcde"]"""
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                item(0, "abcdg".."abcdz")
+            }
+        }
+        expect("[0]: JSON value not in range - \"abcde\"") { exception.message }
+    }
+
+    @Test fun `should fail on test for value as member of a collection of other class`() {
+        val json = "\"C\""
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                value(setOf('C'))
+            }
+        }
+        expect("Can't perform test using collection of kotlin.Char?") { exception.message }
+    }
+
+    @Test fun `should fail on test for property as member of a collection of other class`() {
+        val json = """{"abc":"C"}"""
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                property("abc", setOf('C'))
+            }
+        }
+        expect("abc: Can't perform test using collection of kotlin.Char?") { exception.message }
+    }
+
+    @Test fun `should fail on test for array item as member of a collection of other class`() {
+        val json = """["C"]"""
+        val exception = assertFailsWith<AssertionError> {
+            expectJSON(json) {
+                item(0, setOf('C'))
+            }
+        }
+        expect("[0]: Can't perform test using collection of kotlin.Char?") { exception.message }
+    }
+
 }