+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 - `JSONExpect`: improved error messages
 - `JSONExpect`: all tests now strongly-typed
+- `JSONExpect`: added ability to specify value as a collection
+- `JSONExpect`: added ability to specify value as an `IntRange`, `LongRange` or `ClosedRange`
 
 ## [0.2] - 2020-04-13
 ### Changed